+++ compiler/llvm-backend/src/codegen/structs.rs
@@ -1,4 +1,3 @@
-
 use ast_types::*;
 use inkwell::types::*;
 use inkwell::values::*;